tickets/src/routes/events/listeners/__tests__/orderCreatedListener.test.ts
import mongoose from 'mongoose';
import { Message } from 'node-nats-streaming';
import { OrderCreatedListener } from "../orderCreatedListener";
import { natsWrapper } from "../../../../natsWrapper";
import { Ticket } from "../../../../models/ticket";
import { OrderCreatedEvent, OrderStatus } from "@phticketing/common";

const setup = async () => {
    // create an instance of the listener
    const listener = new OrderCreatedListener(natsWrapper.client);

    // create and save ticket
    const ticket = Ticket.build({
        title: 'Concert',
        price: 99,
        userId: 'fadfadsf'
    });
    await ticket.save();

    // create fake data object
    const data: OrderCreatedEvent['data'] = {
        id: mongoose.Types.ObjectId().toHexString(),
        status: OrderStatus.Created,
        version: 0,
        userId: 'fasdfa',
        expiresAt: 'fadfa',
        ticket: {
            id: ticket.id,
            price: ticket.price
        }
    }

    // create fake message object
    // @ts-ignore
    const msg: Message = {
        ack: jest.fn()
    }

    return { listener, data, ticket, msg };
};

it('set the userId of the ticket', async () => {
    const { listener, data, ticket, msg } = await setup();

    await listener.onMessage(data, msg);

    const updatedTicket = await Ticket.findById(ticket.id);

    expect(updatedTicket!.orderId).toEqual(data.id);
});

it('acks the message', async () => {
    const { listener, data, msg } = await setup();

    await listener.onMessage(data, msg);

    expect(msg.ack).toHaveBeenCalled();
});

it('publishes a ticket updated event', async () => {
    const { listener, data, msg } = await setup();

    await listener.onMessage(data, msg);

    expect(natsWrapper.client.publish).toHaveBeenCalled();

    const ticketUpdatedData = JSON.parse((natsWrapper.client.publish as jest.Mock).mock.calls[0][1]);

    expect(data.id).toEqual(ticketUpdatedData.orderId);
});